requirements.txt
fastapi==0.115.0
uvicorn==0.23.2
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.0.3
mysql-connector-python==8.2.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
starlette==0.38.6
typing-extensions==4.15.0
annotated-types==0.7.0